backend/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
Django==4.1.3
django-extensions==3.2.1
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
hashids==1.3.1
idna==3.4
Pillow==9.3.0
psycopg2-binary
pycparser==2.21
pydot==1.4.2
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
whitenoise==6.2.0
